requirements.txt
matplotlib
numpy
Pillow
opencv-python
scipy